+++ src/js/users/components/__tests__/Group.test.jsx
@@ -1,7 +1,5 @@
 import { UserGroup } from "../Group";
-import React from "react";
 import { screen } from "@testing-library/react";
-import userEvent from "@testing-library/user-event";
 
 describe("<UserGroup />", () => {
     let props;
